refactor portfolio projected cashflow router
